+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"D:\\Полина\\work\\full-screen-carousel-react\\src\\index.js":"1","D:\\Полина\\work\\full-screen-carousel-react\\src\\App.js":"2","D:\\Полина\\work\\full-screen-carousel-react\\src\\serviceWorker.js":"3","D:\\Полина\\work\\full-screen-carousel-react\\src\\Slider.js":"4"},{"size":503,"mtime":499162500000,"results":"5","hashOfConfig":"6"},{"size":203,"mtime":1607864659490,"results":"7","hashOfConfig":"6"},{"size":5086,"mtime":499162500000,"results":"8","hashOfConfig":"6"},{"size":867,"mtime":1607861338147,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"81fld2",{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"D:\\Полина\\work\\full-screen-carousel-react\\src\\index.js",[],"D:\\Полина\\work\\full-screen-carousel-react\\src\\App.js",[],"D:\\Полина\\work\\full-screen-carousel-react\\src\\serviceWorker.js",[],["19","20"],"D:\\Полина\\work\\full-screen-carousel-react\\src\\Slider.js",[],{"ruleId":"21","replacedBy":"22"},{"ruleId":"23","replacedBy":"24"},"no-native-reassign",["25"],"no-negated-in-lhs",["26"],"no-global-assign","no-unsafe-negation"]
+[{"D:\\Полина\\work\\full-screen-carousel-react\\src\\index.js":"1","D:\\Полина\\work\\full-screen-carousel-react\\src\\App.js":"2","D:\\Полина\\work\\full-screen-carousel-react\\src\\serviceWorker.js":"3","D:\\Полина\\work\\full-screen-carousel-react\\src\\Carousel.js":"4"},{"size":503,"mtime":499162500000,"results":"5","hashOfConfig":"6"},{"size":203,"mtime":1607864659490,"results":"7","hashOfConfig":"6"},{"size":5086,"mtime":499162500000,"results":"8","hashOfConfig":"6"},{"size":867,"mtime":1607861338147,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"81fld2",{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"D:\\Полина\\work\\full-screen-carousel-react\\src\\index.js",[],"D:\\Полина\\work\\full-screen-carousel-react\\src\\App.js",[],"D:\\Полина\\work\\full-screen-carousel-react\\src\\serviceWorker.js",[],["19","20"],"D:\\Полина\\work\\full-screen-carousel-react\\src\\Carousel.js",[],{"ruleId":"21","replacedBy":"22"},{"ruleId":"23","replacedBy":"24"},"no-native-reassign",["25"],"no-negated-in-lhs",["26"],"no-global-assign","no-unsafe-negation"]
