+++ src/App.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import Slider from "./Slider";
+import Slider from "./Carousel";
 import './App.css';
 
 function App() {
