+++ src/Carousel.css
@@ -1,0 +1,51 @@
+.carousel {
+  width: 100vw;
+  height: 100vh;
+  overflow: hidden;
+  position: relative;
+}
+
+.carousel__list {
+  outline: seagreen 1px solid;
+  width: 10000000000000vw;
+  height: 100vh;
+  display: flex;
+  list-style-type: none;
+  padding: 0;
+  position: absolute;
+  transition: 1s;
+  margin: 0 auto auto 0;
+}
+
+.carousel__item {
+  width: 100vw;
+  height: 100vh;
+  display: inline;
+  outline: deeppink 1px solid;
+}
+
+.slide {
+  border: deeppink 1px solid;
+  box-sizing: border-box;
+  min-width: 100%;
+  height: 80vh;
+  width: 100vw;
+  display: inline;
+}
+
+.button {
+  position: absolute;
+  top: 50%;
+  transform: translateY(-50%);
+  width: 15%;
+  height: 80vh;
+  background: none;
+}
+
+.button-left {
+  left: 0;
+}
+
+.button-right {
+  right: 0;
+}
