--- src/Slider.css
@@ -1,37 +1,0 @@
-.slider {
-  border: seagreen 1px solid;
-  width: 100%;
-  height: 100vh;
-  box-sizing: border-box;
-  margin: 0;
-  padding: 0;
-  display: flex;
-  position: relative;
-  align-items: center;
-  overflow: hidden;
-}
-
-.slide {
-  border: deeppink 1px solid;
-  box-sizing: border-box;
-  min-width: 100%;
-  height: 80vh;
-  transition: 1s;
-}
-
-.button {
-  position: absolute;
-  top: 50%;
-  transform: translateY(-50%);
-  width: 15%;
-  height: 80vh;
-  background: none;
-}
-
-.button-left {
-  left: 0;
-}
-
-.button-right {
-  right: 0;
-}
